fixing mask projection

+++ util/utilIO.cpp
@@ -110,11 +110,9 @@ cv::Mat ImgIO::projChngMaskTo3D(const cv::Mat &chngMask, const vcg::Shot<float> 
   cam2_fmat = cam2_intr*cam2_Rt;
 
   std::cout<<"Rt 1:\n"<<cam1_Rt<<"\n intr 1:\n"<<cam1_intr<<std::endl;
-  //  std::cout<<"P matrix 1:"<<cam1_fmat<<"\nP matrix 2:"<<cam2_fmat<<std::endl;
   
-  cv::perspectiveTransform(cam1_points, cam2_points, H);
-  
-  cv::Mat pnts3D(4, cam1_points.size(), CV_64FC4);
+  cv::perspectiveTransform(cam1_points, cam2_points, H);  
+  cv::Mat pnts3D(4, cam1_points.size(), CV_32FC4);
 
   cv::triangulatePoints(cam1_fmat, cam2_fmat, cv::Mat(cam1_points).reshape(1,2), cv::Mat(cam2_points).reshape(1,2), pnts3D);
